% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_col_names.R
\name{get_col_names}
\alias{get_col_names}
\title{get_col_names}
\usage{
get_col_names(page_number)
}
\arguments{
\item{page_number}{page number of pdf}
}
\value{
column names for given data frame
}
\description{
get_col_names
}
