% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_file.R
\name{get_example_file}
\alias{get_example_file}
\title{get_example_file}
\usage{
get_example_file(file = NULL)
}
\arguments{
\item{file}{should be null}
}
\value{
path to example file
}
\description{
get_example_file
}
