% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_raw_text.R
\name{parse_raw_text}
\alias{parse_raw_text}
\title{parse_raw_text}
\usage{
parse_raw_text(data, page_number)
}
\arguments{
\item{data}{raw text from selected page as a 1 column data.frame}

\item{page_number}{page number of selected text}
}
\value{
a roughly cleaned data.frame due to non standard inputs, it will need to be further processed
}
\description{
parse_raw_text
}
