% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pdf_text.R
\name{get_pdf_page}
\alias{get_pdf_page}
\title{get_pdf_page}
\usage{
get_pdf_page(pdf, page_num)
}
\arguments{
\item{pdf}{path to pdf that contains a text layer}

\item{page_num}{page number you want text from}
}
\value{
raw text from a page of a pdf file
}
\description{
get_pdf_page
}
